Code cleanup (#277)

+++ langstream-k8s-deployer/langstream-k8s-deployer-core/src/main/java/ai/langstream/deployer/k8s/util/KubeUtil.java
@@ -88,9 +88,9 @@ public class KubeUtil {
             return false;
         }
 
-        final int replicas = status.getReplicas().intValue();
-        final int ready = status.getReadyReplicas().intValue();
-        final int updated = status.getUpdatedReplicas().intValue();
+        final int replicas = status.getReplicas();
+        final int ready = status.getReadyReplicas();
+        final int updated = status.getUpdatedReplicas();
         return replicas == ready && updated == ready;
     }
